+++ test/unit/test_cli.py
@@ -1,23 +1,23 @@
 from unittest.mock import patch
 
 import pytest
-from pullbug.cli import PullBug
+from pullbug.cli import PullBugCli
 
 
 @patch('pullbug.cli.LOGGER')
 def test_throw_missing_error(mock_logger):
     with pytest.raises(ValueError):
-        PullBug.throw_missing_error('GITHUB_TOKEN')
+        PullBugCli.throw_missing_error('GITHUB_TOKEN')
 
     mock_logger.critical.assert_called_once_with('No GITHUB_TOKEN set. Please correct and try again.')
 
 
 @patch('pullbug.cli.GITHUB_TOKEN', '123')
-@patch('pullbug.cli.PullBug.run_missing_checks')
+@patch('pullbug.cli.PullBugCli.run_missing_checks')
 @patch('pullbug.github_bug.GithubBug.run')
 @patch('pullbug.cli.LOGGER')
 def test_run_with_github_arg(mock_logger, mock_github, mock_missing_checks):
-    PullBug.run(True, False, False, False, False, False, 'mock-owner', 'open', 'orgs', 'opened', 'all')
+    PullBugCli.run(True, False, False, False, False, False, 'mock-owner', 'open', 'orgs', 'opened', 'all')
 
     mock_missing_checks.assert_called_once_with(True, False, False, False, False)
     mock_github.assert_called_once()
@@ -29,7 +29,7 @@ def test_run_with_github_arg(mock_logger, mock_github, mock_missing_checks):
 def test_run_missing_checks_no_discord_webhook_url(mock_logger):
     message = 'No DISCORD_WEBHOOK_URL set. Please correct and try again.'
     with pytest.raises(ValueError):
-        PullBug.run_missing_checks(True, False, True, False, False)
+        PullBugCli.run_missing_checks(True, False, True, False, False)
 
     mock_logger.critical.assert_called_once_with(message)
 
@@ -38,7 +38,7 @@ def test_run_missing_checks_no_discord_webhook_url(mock_logger):
 def test_run_missing_checks_no_github_token(mock_logger):
     message = 'No GITHUB_TOKEN set. Please correct and try again.'
     with pytest.raises(ValueError):
-        PullBug.run_missing_checks(True, False, False, False, False)
+        PullBugCli.run_missing_checks(True, False, False, False, False)
 
     mock_logger.critical.assert_called_once_with(message)
 
@@ -48,7 +48,7 @@ def test_run_missing_checks_no_github_token(mock_logger):
 def test_run_missing_checks_no_slack_bot_token(mock_logger):
     message = 'No SLACK_BOT_TOKEN set. Please correct and try again.'
     with pytest.raises(ValueError):
-        PullBug.run_missing_checks(True, False, False, True, False)
+        PullBugCli.run_missing_checks(True, False, False, True, False)
 
     mock_logger.critical.assert_called_once_with(message)
 
@@ -59,7 +59,7 @@ def test_run_missing_checks_no_slack_bot_token(mock_logger):
 def test_run_missing_checks_no_slack_channel(mock_logger):
     message = 'No SLACK_CHANNEL set. Please correct and try again.'
     with pytest.raises(ValueError):
-        PullBug.run_missing_checks(True, False, False, True, False)
+        PullBugCli.run_missing_checks(True, False, False, True, False)
 
     mock_logger.critical.assert_called_once_with(message)
 
@@ -69,7 +69,7 @@ def test_run_missing_checks_no_slack_channel(mock_logger):
 def test_run_missing_checks_no_rocket_chat_url(mock_logger):
     message = 'No ROCKET_CHAT_URL set. Please correct and try again.'
     with pytest.raises(ValueError):
-        PullBug.run_missing_checks(True, False, False, False, True)
+        PullBugCli.run_missing_checks(True, False, False, False, True)
 
     mock_logger.critical.assert_called_once_with(message)
 
@@ -80,6 +80,6 @@ def test_run_missing_checks_no_rocket_chat_url(mock_logger):
 @patch('pullbug.cli.GITHUB_TOKEN', '123')
 @patch('pullbug.cli.LOGGER')
 def test_run_missing_checks_no_errors(mock_logger):
-    PullBug.run_missing_checks(True, False, False, False, True)
+    PullBugCli.run_missing_checks(True, False, False, False, True)
 
     mock_logger.critical.assert_not_called()