+++ test/unit/test_github_bug.py
@@ -79,12 +79,12 @@ def test_run_with_rocketchat(mock_logger, mock_get_repos, mock_pull_request, moc
 @patch('pullbug.github_bug.GITHUB_HEADERS')
 @patch('pullbug.github_bug.LOGGER')
 @patch('requests.get')
-def test_get_repos_success(mock_request, mock_logger, mock_headers, _mock_user, _mock_github_context):
+def test_get_repos_success(mock_request, mock_logger, mock_headers, mock_user, mock_github_context):
     # TODO: Mock this request better and assert additional values
-    GithubBug.get_repos(_mock_user, _mock_github_context)
+    GithubBug.get_repos(mock_user, mock_github_context)
 
     mock_request.assert_called_once_with(
-        f'https://api.github.com/{_mock_github_context}/{_mock_user}/repos?per_page=100', headers=mock_headers
+        f'https://api.github.com/{mock_github_context}/{mock_user}/repos?per_page=100', headers=mock_headers
     )
     assert mock_logger.info.call_count == 2
 
@@ -93,20 +93,20 @@ def test_get_repos_success(mock_request, mock_logger, mock_headers, _mock_user, 
 @patch('pullbug.github_bug.GITHUB_HEADERS')
 @patch('pullbug.github_bug.LOGGER')
 @patch('requests.get', return_value=MockResponse(text='Not Found'))
-def test_get_repos_value_exception(mock_request, mock_logger, mock_headers, _mock_user, _mock_github_context):
+def test_get_repos_value_exception(mock_request, mock_logger, mock_headers, mock_user, mock_github_context):
     with pytest.raises(ValueError):
-        GithubBug.get_repos(_mock_user, _mock_github_context)
+        GithubBug.get_repos(mock_user, mock_github_context)
 
     mock_logger.error.assert_called_once_with(
-        f'Could not retrieve GitHub repos due to bad parameter: {_mock_user} | {_mock_github_context}.'
+        f'Could not retrieve GitHub repos due to bad parameter: {mock_user} | {mock_github_context}.'
     )
 
 
 @patch('pullbug.github_bug.LOGGER')
 @patch('requests.get', side_effect=requests.exceptions.RequestException('mock-error'))
-def test_get_repos_requests_exception(mock_request, mock_logger, _mock_user, _mock_github_context):
+def test_get_repos_requests_exception(mock_request, mock_logger, mock_user, mock_github_context):
     with pytest.raises(requests.exceptions.RequestException):
-        GithubBug.get_repos(_mock_user, _mock_github_context)
+        GithubBug.get_repos(mock_user, mock_github_context)
 
     mock_logger.error.assert_called_once_with('Could not retrieve GitHub repos: mock-error')
 
@@ -115,13 +115,13 @@ def test_get_repos_requests_exception(mock_request, mock_logger, _mock_user, _mo
 @patch('pullbug.github_bug.GITHUB_HEADERS')
 @patch('pullbug.github_bug.LOGGER')
 @patch('requests.get')
-def test_get_pull_requests_success(mock_request, mock_logger, mock_headers, _mock_repo, _mock_github_state, _mock_user):
+def test_get_pull_requests_success(mock_request, mock_logger, mock_headers, mock_repo, mock_github_state, mock_user):
     # TODO: Mock this request better and assert additional values
-    mock_repos = [_mock_repo]
-    result = GithubBug.get_pull_requests(mock_repos, _mock_user, _mock_github_state)
+    mock_repos = [mock_repo]
+    result = GithubBug.get_pull_requests(mock_repos, mock_user, mock_github_state)
 
     mock_request.assert_called_once_with(
-        f'https://api.github.com/repos/{_mock_user}/{_mock_repo["name"]}/pulls?state={_mock_github_state}&per_page=100',
+        f'https://api.github.com/repos/{mock_user}/{mock_repo["name"]}/pulls?state={mock_github_state}&per_page=100',
         headers=mock_headers,
     )
     assert mock_logger.info.call_count == 2
@@ -134,14 +134,14 @@ def test_get_pull_requests_success(mock_request, mock_logger, mock_headers, _moc
 @patch('pullbug.github_bug.LOGGER')
 @patch('requests.get', return_value=None)
 def test_get_pull_requests_success_no_pull_requests(
-    mock_request, mock_logger, mock_headers, _mock_repo, _mock_github_state, _mock_user
+    mock_request, mock_logger, mock_headers, mock_repo, mock_github_state, mock_user
 ):  # noqa
     # TODO: Mock this request better and assert additional values
-    mock_repos = [_mock_repo]
-    result = GithubBug.get_pull_requests(mock_repos, _mock_user, _mock_github_state)
+    mock_repos = [mock_repo]
+    result = GithubBug.get_pull_requests(mock_repos, mock_user, mock_github_state)
 
     mock_request.assert_called_once_with(
-        f'https://api.github.com/repos/{_mock_user}/{_mock_repo["name"]}/pulls?state={_mock_github_state}&per_page=100',
+        f'https://api.github.com/repos/{mock_user}/{mock_repo["name"]}/pulls?state={mock_github_state}&per_page=100',
         headers=mock_headers,
     )
     assert isinstance(result, list)
@@ -150,41 +150,41 @@ def test_get_pull_requests_success_no_pull_requests(
 
 @patch('pullbug.github_bug.LOGGER')
 @patch('requests.get', side_effect=requests.exceptions.RequestException('mock-error'))
-def test_get_pull_requests_request_exception(mock_request, mock_logger, _mock_repo, _mock_user, _mock_github_state):
-    mock_repos = [_mock_repo]
+def test_get_pull_requests_request_exception(mock_request, mock_logger, mock_repo, mock_user, mock_github_state):
+    mock_repos = [mock_repo]
     with pytest.raises(requests.exceptions.RequestException):
-        GithubBug.get_pull_requests(mock_repos, _mock_user, _mock_github_state)
+        GithubBug.get_pull_requests(mock_repos, mock_user, mock_github_state)
 
     mock_logger.error.assert_called_once_with(
-        f'Could not retrieve GitHub pull requests for {_mock_repo["name"]}: mock-error'
+        f'Could not retrieve GitHub pull requests for {mock_repo["name"]}: mock-error'
     )
 
 
 @patch('pullbug.github_bug.LOGGER')
 @patch('requests.get', side_effect=TypeError('mock-error'))
-def test_get_pull_requests_type_error_exception(mock_request, mock_logger, _mock_repo, _mock_user, _mock_github_state):
-    mock_repos = [_mock_repo]
+def test_get_pull_requests_type_error_exception(mock_request, mock_logger, mock_repo, mock_user, mock_github_state):
+    mock_repos = [mock_repo]
     with pytest.raises(TypeError):
-        GithubBug.get_pull_requests(mock_repos, _mock_user, _mock_github_state)
+        GithubBug.get_pull_requests(mock_repos, mock_user, mock_github_state)
 
     mock_logger.error.assert_called_once_with(
-        f'Could not retrieve GitHub pull requests due to bad parameter: {_mock_user} | {_mock_github_state}.'
+        f'Could not retrieve GitHub pull requests due to bad parameter: {mock_user} | {mock_github_state}.'
     )
 
 
 @patch('pullbug.github_bug.Messages.prepare_github_message', return_value=[['mock-message'], ['mock-message']])
-def test_iterate_pull_requests_wip_title(mock_prepare_message, _mock_pull_request):
-    _mock_pull_request['title'] = 'wip: mock-pull-request'
-    mock_pull_requests = [_mock_pull_request]
+def test_iterate_pull_requests_wip_title(mock_prepare_message, mock_pull_request):
+    mock_pull_request['title'] = 'wip: mock-pull-request'
+    mock_pull_requests = [mock_pull_request]
     GithubBug.iterate_pull_requests(mock_pull_requests, True, False, False, False)
 
     mock_prepare_message.assert_called_once()
 
 
 @patch('pullbug.github_bug.Messages.prepare_github_message')
-def test_iterate_pull_requests_wip_setting_absent(mock_prepare_message, _mock_pull_request):
-    _mock_pull_request['title'] = 'wip: mock-pull-request'
-    mock_pull_requests = [_mock_pull_request]
+def test_iterate_pull_requests_wip_setting_absent(mock_prepare_message, mock_pull_request):
+    mock_pull_request['title'] = 'wip: mock-pull-request'
+    mock_pull_requests = [mock_pull_request]
     GithubBug.iterate_pull_requests(mock_pull_requests, False, False, False, False)
 
     mock_prepare_message.assert_not_called()