+++ CHANGELOG.md
@@ -13,9 +13,10 @@
 ### Other Changes
 
 * Added new `--repos` flag to filter repos you want pull requests for (closes #23)
-* Added new `--issues` flag to bug GitHub for issues instead of pull requests. No includes `--pulls` flag to help differentiate
+* Added new `--issues` flag to bug GitHub for issues instead of pull requests. No includes `--pulls` flag to help differentiate (closes #19)
 * Switched from raw API calls to PyGithub which allows us to properly handle pagination (closes #14)
 * Replace all `classmethods` with instance or static methods
+* Various bug fixes and code refactor
 
 ## v2.4.0 (2021-09-20)